requirements.txt
transformers>=4.43.0
accelerate>=0.33.0
peft>=0.12.0
trl>=0.9.4
bitsandbytes>=0.43.0
datasets>=2.20.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
torchaudio>=2.3.0
soundfile>=0.12.1
pandas>=2.2.2
numpy>=1.26.4
pyyaml>=6.0.1
sqlite-utils>=3.36
jiwer>=3.0.3
tqdm>=4.66.4
matplotlib>=3.8.0
orjson>=3.10.7
jsonlines>=4.0.0
python-dateutil>=2.9.0
